Updated based on David's comments..
