+++ bundle/src/main/java/com/adobe/acs/commons/notification/inbox/AEMInboxNotificationDetails.java
@@ -1,88 +1,30 @@
-/*
- * #%L
- * ACS AEM Commons Bundle
- * %%
- * Copyright (C) 2013 Adobe
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
 package com.adobe.acs.commons.notification.inbox;
 
 import java.io.Serializable;
 
-public class AEMInboxNotificationDetails implements Serializable {
+public interface AEMInboxNotificationDetails extends Serializable {
+    
+    public String getTitle();
 
-    private static final long serialVersionUID = -5976192100927192675L;
+    public void setTitle(String title);
 
-    private String title;
+    public String getContentPath();
 
-    private String contentPath;
+    public void setContentPath(String contentPath);
 
-    private String assignee;
+    public String getAssignee();
 
-    private String messageDetails;
+    public void setAssignee(String assignee);
 
-    private String[] notificationAction;
+    public String getMessageDetails();
 
-    private String instructions;
+    public void setMessageDetails(String messageDetails);
 
-    public String getTitle() {
-        return title;
-    }
+    public String[] getNotificationAction();
 
-    public void setTitle(String title) {
-        this.title = title;
-    }
+    public void setNotificationAction(String[] notificationAction);
 
-    public String getContentPath() {
-        return contentPath;
-    }
+    public String getInstructions();
 
-    public void setContentPath(String contentPath) {
-        this.contentPath = contentPath;
-    }
-
-    public String getAssignee() {
-        return assignee;
-    }
-
-    public void setAssignee(String assignee) {
-        this.assignee = assignee;
-    }
-
-    public String getMessageDetails() {
-        return messageDetails;
-    }
-
-    public void setMessageDetails(String messageDetails) {
-        this.messageDetails = messageDetails;
-    }
-
-    public String[] getNotificationAction() {
-        return notificationAction;
-    }
-
-    public void setNotificationAction(String[] notificationAction) {
-        this.notificationAction = notificationAction;
-    }
-
-    public String getInstructions() {
-        return instructions;
-    }
-
-    public void setInstructions(String instructions) {
-        this.instructions = instructions;
-    }
-
+    public void setInstructions(String instructions);
 }
