+++ bundle/src/main/java/com/adobe/acs/commons/notification/inbox/impl/AEMInboxNotificationDetailsImpl.java
@@ -1,0 +1,88 @@
+/*
+ * #%L
+ * ACS AEM Commons Bundle
+ * %%
+ * Copyright (C) 2013 Adobe
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package com.adobe.acs.commons.notification.inbox.impl;
+
+import com.adobe.acs.commons.notification.inbox.AEMInboxNotificationDetails;
+
+public class AEMInboxNotificationDetailsImpl implements AEMInboxNotificationDetails {
+
+    private static final long serialVersionUID = -5976192100927192675L;
+
+    private String title;
+
+    private String contentPath;
+
+    private String assignee;
+
+    private String messageDetails;
+
+    private String[] notificationAction;
+
+    private String instructions;
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getContentPath() {
+        return contentPath;
+    }
+
+    public void setContentPath(String contentPath) {
+        this.contentPath = contentPath;
+    }
+
+    public String getAssignee() {
+        return assignee;
+    }
+
+    public void setAssignee(String assignee) {
+        this.assignee = assignee;
+    }
+
+    public String getMessageDetails() {
+        return messageDetails;
+    }
+
+    public void setMessageDetails(String messageDetails) {
+        this.messageDetails = messageDetails;
+    }
+
+    public String[] getNotificationAction() {
+        return notificationAction;
+    }
+
+    public void setNotificationAction(String[] notificationAction) {
+        this.notificationAction = notificationAction;
+    }
+
+    public String getInstructions() {
+        return instructions;
+    }
+
+    public void setInstructions(String instructions) {
+        this.instructions = instructions;
+    }
+
+}
